+++ modules/core/inventory/src/AssetInventoryInterface.php
@@ -18,10 +18,13 @@ interface AssetInventoryInterface {
    *   The quantity measure of the inventory. See quantity_measures().
    * @param int $units
    *   The quantity units of the inventory (term ID).
+   * @param int|null $timestamp
+   *   Include logs with a timestamp less than or equal to this.
+   *   If this is NULL (default), the current time will be used.
    *
    * @return array
    *   Returns an array of asset inventory information.
    */
-  public function getInventory(AssetInterface $asset, string $measure = '', int $units = 0): array;
+  public function getInventory(AssetInterface $asset, string $measure = '', int $units = 0, $timestamp = NULL): array;
 
 }